Fix: Change diff_mult calculations, default momentum values, and beta1.
The current default momentums were tested on simple synthetic tests and LoRA/DoKr training for SDXL, for approximately ~2000 steps. If you use many steps (more than 20,000,) the gradient (and resulting step distance) may become too large. In this case, use the alternative momentum beta and lambda. The alternative momentum may be slightly less smooth, but won't be amplified nearly as much.

Alternatively, momentum_beta may be set to 0.9999, and if changed, momentum_lambda should also be set to 0.25.

There was previously not much learning due to improper clamping and diff_mult. These changes ought to result in a smoother gradient and should be rather robust to noise once again.